requirements.txt
Flask==2.2.5
pytest==7.1.3
requests==2.28.0
# pact-python==3.0.0b0
pact-python @ https://github.com/YOU54F/pact-python/archive/3.0.0b0.tar.gz